- [fix] [WEBMIS-633] Сброс направления в отделение - изменена логика, теперь отсутствующие в анализе свойства действия игнорируются (не сбрасываются);

+++ core/business-tier/src/main/scala/ru/korus/tmis/core/patient/AppealBean.scala
@@ -185,17 +185,17 @@ with CAPids{
                                          authData)
         list = actionPropertyBean.getActionPropertiesByActionId(temp.getId.intValue).keySet.toList
 
-        var list2 = actionPropertyTypeBean.getActionPropertyTypesByActionTypeId(i18n("db.actionType.hospitalization.primary").toInt)
+        val list2 = actionPropertyTypeBean.getActionPropertyTypesByActionTypeId(i18n("db.actionType.hospitalization.primary").toInt)
                                           .toList
                                           .filter(p => {
-          val sadasd = list.filter(pp => pp.asInstanceOf[ActionProperty].getType.getId == p.getId)
-          if (sadasd ==null || sadasd.size == 0) true else false
+          val filtred = list.filter(pp => pp.asInstanceOf[ActionProperty].getType.getId == p.getId)
+          if (filtred == null || filtred.size == 0) true else false
         })
         list2.foreach(ff => { //создание недостающих акшен пропертей
-          val res = actionPropertyBean.createActionProperty(action, ff.asInstanceOf[ActionPropertyType].getId.intValue(), authData)
+          val res = actionPropertyBean.createActionProperty(action, ff.getId.intValue(), authData)
           em.persist(res)
         })
-        //пересобираем листь акшенПропертей
+        //пересобираем лист акшенПропертей
         list = actionPropertyBean.getActionPropertiesByActionId(temp.getId.intValue).keySet.toList
       }
 
@@ -225,45 +225,47 @@ with CAPids{
           entities = entities + ap
 
         val values = this.getValueByCase(ap.getType.getId.intValue(), appealData, authData)
-        values.size match {
-          case 0 => {
-            if (flgCreate) {
-              //В случае, если на приходит значение для ActionProperty, то записываем значение по умолчанию.
-              val defValue = ap.getType.getDefaultValue
-              if (defValue!=null && !defValue.trim.isEmpty) {
-                val apv = actionPropertyBean.setActionPropertyValue(ap, defValue, 0)
+        if (values!=null){
+          values.size match {
+            case 0 => {
+              if (flgCreate) {
+                //В случае, если на приходит значение для ActionProperty, то записываем значение по умолчанию.
+                val defValue = ap.getType.getDefaultValue
+                if (defValue!=null && !defValue.trim.isEmpty) {
+                  val apv = actionPropertyBean.setActionPropertyValue(ap, defValue, 0)
+                  if (apv!=null)
+                    entities = entities + apv.unwrap
+                }
+              } else { //Если пришел пустой список, а старые значения есть, то зачистим их
+                val apvs = actionPropertyBean.getActionPropertyValue(ap)
+                if (apvs!=null && apvs.size()>0) {
+                  for(i <- 0 until apvs.size){
+                    var apv = apvs(i).unwrap()
+                    apv = em.merge(apv)
+                    em.remove(apv)
+                  }
+                }
+              }
+            }
+            case _ => {
+              if(ap.getType.getIsVector) { //Если вектор, то сперва зачищаем старый список
+              val apvs = actionPropertyBean.getActionPropertyValue(ap)
+                if (apvs!=null && apvs.size()>values.size){
+                  for(i <- values.size to apvs.size-1) { //если новых значений меньше тем старых, то хвост зачистим
+                    var apv = apvs(i).unwrap()
+                    apv = em.merge(apv)
+                    em.remove(apv)
+                  }
+                }
+              }
+              var it = 0
+              values.foreach(value => {
+                val apv = actionPropertyBean.setActionPropertyValue(ap, value, it)
                 if (apv!=null)
                   entities = entities + apv.unwrap
-              }
-            } else { //Если пришел пустой список, а старые значения есть, то зачистим их
-              val apvs = actionPropertyBean.getActionPropertyValue(ap)
-              if (apvs!=null && apvs.size()>0) {
-                for(i <- 0 until apvs.size){
-                  var apv = apvs(i).unwrap()
-                  apv = em.merge(apv)
-                  em.remove(apv)
-                }
-              }
+                it = it + 1
+              })
             }
-          }
-          case _ => {
-            if(ap.getType.getIsVector) { //Если вектор, то сперва зачищаем старый список
-            val apvs = actionPropertyBean.getActionPropertyValue(ap)
-              if (apvs!=null && apvs.size()>values.size){
-                for(i <- values.size to apvs.size-1) { //если новых значений меньше тем старых, то хвост зачистим
-                  var apv = apvs(i).unwrap()
-                  apv = em.merge(apv)
-                  em.remove(apv)
-                }
-              }
-            }
-            var it = 0
-            values.foreach(value => {
-              val apv = actionPropertyBean.setActionPropertyValue(ap, value, it)
-              if (apv!=null)
-                entities = entities + apv.unwrap
-              it = it + 1
-            })
           }
         }
       })
@@ -607,7 +609,7 @@ with CAPids{
 
   private def AnyToSetOfString(that: AnyRef, sec: String): Set[String] = {
     if(that==null)
-      return Set.empty[String]
+      return null //Set.empty[String]     //В случае если не обрабатываем проперти вернем нулл (чтобы не переписывать значения)
 
     if (that.isInstanceOf[Date]) {
       return Set(ConfigManager.DateFormatter.format(that))
